Aktualizacja serwisu - działa poprawnie

+++ Final_project/virtual_workshop/views.py
@@ -1,9 +1,9 @@
 from django.shortcuts import render, redirect
-from django.views import View
 from .models import Tools, Service
 from django.contrib import messages
 from django.views.decorators.http import require_POST
 from django.http import HttpResponseRedirect, HttpResponse
+from django.utils import timezone
 
 # Create your views here.
 
@@ -29,18 +29,27 @@ def add_tools(request):
         messages.success(request, "Narzędzie zostało dodane!")
         return redirect('tools')
 
+def buy_new_tool(request, tool_id):
+    tool = Tools.objects.get(id=tool_id)
+    tool.quantity += 1
+    tool.save()
+    messages.success(request, f"Ilość narzędzia {tool.name} została zwiększona.")
+    return redirect('tools')
 @require_POST
 def delete_tool(request, tool_id):
     tool = Tools.objects.get(id=tool_id)
-    tool.delete()
+    if tool.quantity > 1:
+        tool.quantity -= 1
+        tool.save()
+    else:
+        tool.delete()
     return redirect('tools')
 
-def tools_status(requset, tool_id):
+def tools_status(request, tool_id):
     tool = Tools.objects.get(Tools, id=tool_id)
     tool.in_job = 'in_job' in request.POST
     tool.in_service = 'in_service' in request.POST
     tool.save()
-    messages.success(request, "Status narzędzia został zaktualizowany.")
     return redirect('tools')
 
 
@@ -55,16 +64,24 @@ def add_tool_to_job(request):
 
 def service(request):
     services = Service.objects.all()
+    for service in services:
+        if service.expected_pickup_date < timezone.now().date() and not service.repaired:
+            service.repaired = True
+            service.save()
     return render(request, 'service.html', {'services': services})
 
 
 def add_tool_to_service(request):
     if request.method == 'GET':
-        tools = Tools.objects.all()
+        tools = Tools.objects.filter(quantity__gt=0)
         return render(request, 'add_tool_to_service.html', {'tools': tools})
     else:
         tool_id = request.POST.get('tool')
         tool = Tools.objects.get(id=tool_id)
+        if tool.quantity > 0:
+            tool.quantity -= 1
+            tool.in_service = True
+            tool.save()
         fault_description = request.POST.get('fault_description')
         expected_pickup_date = request.POST.get('expected_pickup_date')
         service = Service(tool=tool, fault_description=fault_description, expected_pickup_date=expected_pickup_date)
@@ -81,6 +98,10 @@ def repair_tool(request, service_id):
 @require_POST
 def take_from_service(request, service_id):
     service = Service.objects.get(id=service_id)
+    tool = service.tool
+    tool.quantity += 1
+    tool.in_service = False
+    tool.save()
     service.delete()
     return redirect('service')
